+++ src/index/art.rs
@@ -32,22 +32,22 @@ impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ art_sys::Ro
 }
 
 impl<K: index::Key> index::IndexPin<K> for art_sys::RowexRef<'_> {
-    fn get(&mut self, key: &K) -> Option<u32> {
+    fn get(&mut self, key: &K) -> Option<u64> {
         art_sys::RowexRef::get(self, unsafe { core::mem::transmute_copy::<K, u64>(key) });
         None
     }
 
-    fn insert(&mut self, key: K, _: u32) -> Option<u32> {
+    fn insert(&mut self, key: K, _: u64) -> Option<u64> {
         art_sys::RowexRef::insert(self, unsafe { core::mem::transmute_copy::<K, u64>(&key) });
         None
     }
 
-    fn update(&mut self, key: K, value: u32) -> Option<u32> {
+    fn update(&mut self, key: K, value: u64) -> Option<u64> {
         self.insert(key, value);
         None
     }
 
-    fn remove(&mut self, key: K) -> Option<u32> {
+    fn remove(&mut self, key: K) -> Option<u64> {
         art_sys::RowexRef::remove(self, unsafe { core::mem::transmute_copy::<K, u64>(&key) });
         None
     }
