+++ src/index/bz_tree.rs
@@ -1,7 +1,7 @@
 use crate::Index;
 use crate::index;
 
-impl<K: index::Key, H: index::Hasher> Index<K, H> for bztree::BzTree<K, u32> {
+impl<K: index::Key, H: index::Hasher> Index<K, H> for bztree::BzTree<K, u64> {
     type Send<'a> = &'a Self;
 
     fn new(_: &index::Config) -> Self {
@@ -13,7 +13,7 @@ impl<K: index::Key, H: index::Hasher> Index<K, H> for bztree::BzTree<K, u32> {
     }
 }
 
-impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ bztree::BzTree<K, u32> {
+impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ bztree::BzTree<K, u64> {
     type Handle<'a>
         = Self
     where
@@ -24,13 +24,13 @@ impl<K: index::Key, H: index::Hasher> index::IndexSend<K, H> for &'_ bztree::BzT
     }
 }
 
-impl<K: index::Key> index::IndexPin<K> for &'_ bztree::BzTree<K, u32> {
-    fn get(&mut self, key: &K) -> Option<u32> {
+impl<K: index::Key> index::IndexPin<K> for &'_ bztree::BzTree<K, u64> {
+    fn get(&mut self, key: &K) -> Option<u64> {
         let guard = &crossbeam_epoch::pin();
         bztree::BzTree::get(self, key, guard).copied()
     }
 
-    fn insert(&mut self, key: K, value: u32) -> Option<u32> {
+    fn insert(&mut self, key: K, value: u64) -> Option<u64> {
         let guard = &crossbeam_epoch::pin();
         bztree::BzTree::upsert(self, key, value, guard).copied()
     }
@@ -40,7 +40,7 @@ impl<K: index::Key> index::IndexPin<K> for &'_ bztree::BzTree<K, u32> {
         _retry_scan: usize,
         min: &'a K,
         max: &'a K,
-        output: &mut Vec<(K, u32)>,
+        output: &mut Vec<(K, u64)>,
     ) {
         let guard = &crossbeam_epoch::pin();
         output.extend(
